refactor, JDupOption has the logic in the call method
